move non-components out of guui
