+++ packages/guui/index.js
@@ -1,11 +1,8 @@
 // @flow
 
 import CloseButton from './components/CloseButton';
-import Number from './components/Number';
+import Numbers from './components/Numbers';
 import SiteMessage from './components/SiteMessage';
+import { Row as GridRow, Cols as GridCols } from './components/Grid';
 
-export {
-    CloseButton,
-    Number,
-    SiteMessage,
-};
+export { CloseButton, Numbers, SiteMessage, GridRow, GridCols };
