+++ packages/guui/package.json
@@ -9,16 +9,6 @@
     "prepublish": "cd ../.. && make pre-publish-guui",
     "postpublish": "cd ../.. && make post-publish-guui"
   },
-  "files": [
-    "document.js",
-    "grid.js",
-    "htmlTemplate.js",
-    "index.js"
-  ],
-  "directories": {
-    "lib": "./lib",
-    "components": "./components"
-  },
   "license": "Apache-2.0",
   "dependencies": {
     "@guardian/pasteup": "0.1.0-alpha.1",
@@ -33,7 +23,6 @@
     "desvg-loader": "^0.1.0",
     "emotion": "^9.1.1",
     "react": "^16.4.0",
-    "react-dom": "^16.4.0",
     "react-emotion": "^9.1.3",
     "svg-loader": "^0.0.2",
     "to-string-loader": "^1.1.5"
@@ -41,7 +30,6 @@
   "peerDependencies": {
     "emotion": "^9.1.1",
     "react": "^16.4.0",
-    "react-dom": "^16.4.0",
     "react-emotion": "^9.1.3"
   }
 }
