+++ sites/frontend/components/MostViewed.js
@@ -5,7 +5,7 @@ import { Component } from 'react';
 import { headline, textEgyptian } from '@guardian/pasteup/fonts';
 import palette from '@guardian/pasteup/palette';
 import { desktop } from '@guardian/pasteup/breakpoints';
-import Numbers from '@guardian/guui/components/Number';
+import { Numbers } from '@guardian/guui';
 
 const Heading = styled('h2')({
     fontFamily: headline,
