+++ sites/frontend/lib/assets.js
