+++ sites/frontend/pages/Article.js
@@ -3,7 +3,7 @@
 /* eslint-disable react/no-danger */
 
 import styled from 'react-emotion';
-import { Row, Cols } from '@guardian/guui/grid';
+import { GridRow as Row, GridCols as Cols } from '@guardian/guui';
 import { textEgyptian, headline } from '@guardian/pasteup/fonts';
 import palette from '@guardian/pasteup/palette';
 import { clearFix } from '@guardian/pasteup/mixins';
